cleanup imports, move config resolution to separate file
