+++ sgrep_lint/sgrep_types.py
@@ -1,21 +1,5 @@
-import argparse
-import base64
-import collections
-import itertools
-import json
-import os
-import shutil
-import subprocess
-import sys
-import tarfile
-import tempfile
-import time
-import traceback
 from dataclasses import dataclass
 from dataclasses import field
-from datetime import datetime
-from pathlib import Path
-from pathlib import PurePath
 from typing import Any
 from typing import DefaultDict
 from typing import Dict
